apps/api/requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# Validation
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Database
supabase>=2.10.0

# HTTP & Utils
python-dotenv>=1.0.0
httpx>=0.28.0
python-multipart>=0.0.12

# AI
google-genai>=1.10.0

# Security
cryptography>=44.0.0
slowapi>=0.1.9

# Channels (opcional - usado em alguns endpoints)
telethon>=1.38.0
imapclient>=3.0.0
